Implemented CSS code style, extracted inline CSS to separate file
