+++ client/index.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <title>Chat application PGR6301</title>
     <link rel="stylesheet" href="styles/style.css">
-    <link rel="stylesheet" href="styles/form.css">
 </head>
 <body>
 <div id="app"></div>
